+++ src/smpp/smpp_sms.h
@@ -25,7 +25,7 @@ using namespace smsc::sms;
 
 inline Address PduAddress2Address(const PduAddress& source)
 {
-  return Address(source.value.cstr()?strlen(source.value.cstr()):1,
+  return Address(source.value.cstr()?(uint8_t)strlen(source.value.cstr()):1,
                  source.typeOfNumber,
                  source.numberingPlan,
                  source.value.cstr()?source.value.cstr():"0");
@@ -354,7 +354,7 @@ inline void fetchOptionals(SmppOptional& optional,SMS* sms,bool forceDC=false)
     sms->setBinProperty(Tag::SMSC_UNKNOWN_OPTIONALS,optional.get_unknownFields(),optional.size_unknownFields());
   }
 //  if ( optional.has_protocol_id() )
-//    sms->setIntProperty(Tag::SMPP_PROTOCOL_ID,(uint32_t)optional.get_protocol_id());  
+//    sms->setIntProperty(Tag::SMPP_PROTOCOL_ID,(uint32_t)optional.get_protocol_id());
 }
 
 inline bool fetchSmsFromSmppPdu(PduXSm* pdu,SMS* sms,bool forceDC=false)
@@ -372,11 +372,11 @@ inline bool fetchSmsFromSmppPdu(PduXSm* pdu,SMS* sms,bool forceDC=false)
     { // fill address
       PduAddress& source = message.source;
       PduAddress& dest  = message.dest;
-      Address originatingAddr(source.value.cstr()?strlen(source.value.cstr()):1,
+      Address originatingAddr(source.value.cstr()?(uint8_t)strlen(source.value.cstr()):1,
                             source.typeOfNumber,
                             source.numberingPlan,
                             source.value.cstr()?source.value.cstr():"0");
-      Address destinationAddr(dest.value.cstr()?strlen(dest.value.cstr()):1,
+      Address destinationAddr(dest.value.cstr()?(uint8_t)strlen(dest.value.cstr()):1,
                             dest.typeOfNumber,
                             dest.numberingPlan,
                             dest.value.cstr()?dest.value.cstr():"0");/**/
@@ -506,11 +506,11 @@ inline bool fetchSmsFromDataSmPdu(PduDataSm* pdu,SMS* sms,bool forceDC=false)
   { // fill address
     PduAddress& source = data.source;
     PduAddress& dest  = data.dest;
-    Address originatingAddr(source.value.cstr()?strlen(source.value.cstr()):1,
+    Address originatingAddr(source.value.cstr()?(uint8_t)strlen(source.value.cstr()):1,
                           source.typeOfNumber,
                           source.numberingPlan,
                           source.value.cstr()?source.value.cstr():"0");
-    Address destinationAddr(dest.value.cstr()?strlen(dest.value.cstr()):1,
+    Address destinationAddr(dest.value.cstr()?(uint8_t)strlen(dest.value.cstr()):1,
                           dest.typeOfNumber,
                           dest.numberingPlan,
                           dest.value.cstr()?dest.value.cstr():"0");
@@ -607,7 +607,7 @@ inline bool fetchSmsFromDataSmPdu(PduDataSm* pdu,SMS* sms,bool forceDC=false)
   sms->setIntProperty(Tag::SMPP_PRIORITY,0);
   sms->setIntProperty(Tag::SMPP_PROTOCOL_ID,0);
   if ( pdu->optional.has_protocol_id() )
-    sms->setIntProperty(Tag::SMPP_PROTOCOL_ID,(uint32_t)pdu->optional.get_protocol_id());  
+    sms->setIntProperty(Tag::SMPP_PROTOCOL_ID,(uint32_t)pdu->optional.get_protocol_id());
   sms->setIntProperty(Tag::SMPP_REPLACE_IF_PRESENT_FLAG,0);
   sms->setNextTime(0);
   return true;
